Rename workflow-config to config-schemas
